requirements.txt
azure-core==1.36.0
azure-identity==1.25.1
azure-purview-catalog==1.0.0b4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
isodate==0.7.2
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
requests-oauthlib==2.0.0
typing_extensions==4.15.0
urllib3==2.5.0
